Update README.md
 The app displays the current temperature, city name, and weather details like humidity and wind speed. The weather icon changes according to the weather condition of the searched city.

The app has a simple and clean UI with a search box and a search button to initiate the search. In case the user enters an invalid city name, the app displays an error message.

The app uses asynchronous JavaScript to fetch the data from the API and updates the UI dynamically using the fetched data. The app also has an event listener that allows the user to search by pressing the enter key in addition to clicking the search button.

Overall, this weather app is a useful tool for checking the weather details of any city in the world with a few clicks.